ARC-164 Integration Test of Humble Migration (#121)
<!-- Thanks for the contribution, this is awesome. -->

# PR Details
## Description
The velodyne_pointcloud is [not exposing `velodyne_convert_node` in ros2
humble.](https://github.com/ros-drivers/velodyne/tree/humble-devel/velodyne_pointcloud).

We used to use that in the foxy version. However, using
`velodyne_transform_node` works fine, so I switched it.

NOTE: I integration tested with Blue lexus did a full run with rest of
the images in foxy. It localized fine; however, when running the vehicle
was turning aggressively a bit when it starts, but stabilizes seconds
later and was slightly running on the left side of the road despite rviz
pointclouds exactly matching the buildings. I haven't run it without
this image, so it could be just trajectory_follower tuning of blue
lexus. Such behavior will be captured in the full regression testing,
but good to keep in mind.

<!--- Describe your changes in detail -->

## Related GitHub Issue
NA
<!--- This project only accepts pull requests related to open GitHub
issues or Jira Keys -->
<!--- If suggesting a new feature or change, please discuss it in an
issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps
to reproduce -->
<!--- Please DO NOT name partially fixed issues, instead open an issue
specific to this fix -->
<!--- Please link to the issue here: -->

## Related Jira Key
ARC-164
<!-- e.g. CAR-123 -->

## Motivation and Context
ROS2 Humble Upgrade integration testing
<!--- Why is this change required? What problem does it solve? -->

## How Has This Been Tested?
Blue lexus did a full run with rest of the images in foxy.
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Defect fix (non-breaking change that fixes an issue)
- [ ] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that cause existing functionality
to change)

## Checklist:

<!--- Go over all the following points, and put an `x` in all the boxes
that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're
here to help! -->

- [ ] I have added any new packages to the sonar-scanner.properties file
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [ ] I have read the
[**CONTRIBUTING**](https://github.com/usdot-fhwa-stol/carma-platform/blob/develop/Contributing.md)
document.
- [ ] I have added tests to cover my changes.
- [ ] All new and existing tests passed.